доработал дома обработку ARP запросов + похоже разобрался с тредами.

+++ target_board/drv/mpcdrv/linux/mpcdrvlbcCyclone.c
@@ -1337,7 +1337,7 @@ void TDM0_dierction_read ()
 		   memcpy(&dir0_target_arp_nms_sa_addr,out_buf+14,4);
 		   memcpy(&dir0_target_arp_ip_da_addr,out_buf+19,4);
 		   //printk("ARP zaprosi on tdm dir0 0x%x\n\r",dir0_target_arp_ip_da_addr);
-		   ngraf_packet_for_matrica_kommutacii(out_buf ,dannie1200,(u8)dir0_target_arp_ip_da_addr,(u8)dir0_target_arp_nms_sa_addr);   
+		   ngraf_packet_for_matrica_kommutacii(out_buf ,dannie1200,dir0_target_arp_ip_da_addr,dir0_target_arp_nms_sa_addr);   
 	 	   
 	 	  //send to tsec2
 		  //p2020_get_recieve_virttsec_packet_buf(out_buf,dannie1200,2);
@@ -1346,7 +1346,7 @@ void TDM0_dierction_read ()
 	   else
 	   {
 	      // printk("paket po ip hesder\n\r");
-		     ngraf_packet_for_matrica_kommutacii(out_buf ,dannie1200,(u8)dir0_ip_da_addr,0x0000); 
+		     ngraf_packet_for_matrica_kommutacii(out_buf ,dannie1200,dir0_ip_da_addr,0x0000); 
 	   }
 	   
 	   //priznak packeta KY-S 
@@ -1449,7 +1449,7 @@ void TDM1_dierction_read ()
 			   memcpy(&dir1_target_arp_nms_sa_addr,out_buf1+14,4);
 			   memcpy(&dir1_target_arp_ip_da_addr,out_buf1+19,4);
 			   //printk("ARP zaprosi on tdm dir0 0x%x\n\r",dir0_target_arp_ip_da_addr);
-			   ngraf_packet_for_matrica_kommutacii(out_buf1 ,dannie1202,(u8)dir1_target_arp_ip_da_addr,(u8)dir1_target_arp_nms_sa_addr);  
+			   ngraf_packet_for_matrica_kommutacii(out_buf1 ,dannie1202,dir1_target_arp_ip_da_addr,dir1_target_arp_nms_sa_addr);  
 			     //ARP packet for matrica
 		 		 //ngraf_packet_for_matrica_kommutacii(out_buf1 ,dannie1202,0x0806); 		   
 			     //p2020_get_recieve_virttsec_packet_buf(out_buf1,dannie1202,2);
@@ -1457,7 +1457,7 @@ void TDM1_dierction_read ()
 	  	 //priznak ky-s
 	  	 else
 	  	 {
-	  		ngraf_packet_for_matrica_kommutacii(out_buf1 ,dannie1202,(u8)dir1_ip_da_addr,0x0000); 	  	 
+	  		ngraf_packet_for_matrica_kommutacii(out_buf1 ,dannie1202,dir1_ip_da_addr,0x0000); 	  	 
 	  	 }
 		 
 		 
@@ -1569,7 +1569,7 @@ void TDM2_dierction_read ()
 	 	  		   memcpy(&dir2_target_arp_nms_sa_addr,out_buf2+14,4);
 	 	  		   memcpy(&dir2_target_arp_ip_da_addr,out_buf2+19,4);
 				   //printk("ARP zaprosi on tdm dir0 0x%x\n\r",dir0_target_arp_ip_da_addr);
-				   ngraf_packet_for_matrica_kommutacii(out_buf2 ,dannie1204,(u8)dir2_target_arp_ip_da_addr,(u8)dir2_target_arp_nms_sa_addr);  
+				   ngraf_packet_for_matrica_kommutacii(out_buf2 ,dannie1204,dir2_target_arp_ip_da_addr,dir2_target_arp_nms_sa_addr);  
 	 	  		 
 	 	  	     //p2020_get_recieve_virttsec_packet_buf(out_buf2,dannie1204,2);//send to eternet tsec ARP broadcast   
 	 	  	   } 
@@ -1681,13 +1681,13 @@ void TDM3_dierction_read  ()
 		 	  		   memcpy(&dir3_target_arp_nms_sa_addr,out_buf3+14,4);
 		 	  		   memcpy(&dir3_target_arp_ip_da_addr,out_buf3+19,4);
 					   //printk("ARP zaprosi on tdm dir0 0x%x\n\r",dir0_target_arp_ip_da_addr);
-					   ngraf_packet_for_matrica_kommutacii(out_buf3 ,dannie1206,(u8)dir3_target_arp_ip_da_addr,(u8)dir3_target_arp_nms_sa_addr);  
+					   ngraf_packet_for_matrica_kommutacii(out_buf3 ,dannie1206,dir3_target_arp_ip_da_addr,dir3_target_arp_nms_sa_addr);  
 		 	  		 
 		 	  	     //p2020_get_recieve_virttsec_packet_buf(out_buf2,dannie1204,2);//send to eternet tsec ARP broadcast   
 		 	  	   } 
 		 	  	   else
 		 	  	   {   
-		 	  	     ngraf_packet_for_matrica_kommutacii(out_buf3 ,dannie1206,(UINT8)dir3_ip_da_addr,0x0000);  
+		 	  	     ngraf_packet_for_matrica_kommutacii(out_buf3 ,dannie1206,dir3_ip_da_addr,0x0000);  
 		 	  	   }
